Finished disk rotation

+++ app/src/main/java/com/example/zakhariystasenko/retrophone/RetroPhoneView/RotationLimiter.java
@@ -3,9 +3,9 @@ package com.example.zakhariystasenko.retrophone.RetroPhoneView;
 import android.graphics.RectF;
 
 class RotationLimiter {
-    static final int mRotationLimiterAngle = -60;
-    static final int mRotationLimiterImageAngle = 69;
-    static final int mRotationLimiterImageSize = 10;
+    static final float mRotationLimiterAngle = -60;
+    static final float mRotationLimiterImageAngle = 69;
+    static final float mRotationLimiterImageSize = 10;
     RectF mRotationLimiterImage;
 
     RotationLimiter(RectF rotationLimiterImage){
